+++ exomiser-core/src/main/java/de/charite/compbio/exomiser/core/factories/SampleDataFactory.java
@@ -8,15 +8,11 @@ package de.charite.compbio.exomiser.core.factories;
 import de.charite.compbio.exomiser.core.model.SampleData;
 import de.charite.compbio.exomiser.core.model.Gene;
 import de.charite.compbio.exomiser.core.model.VariantEvaluation;
-import de.charite.compbio.exomiser.core.util.VariantAnnotator;
-import jannovar.exception.JannovarException;
-import jannovar.exception.VCFParseException;
 import jannovar.exome.Variant;
 import jannovar.io.VCFReader;
 import jannovar.pedigree.Pedigree;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,10 +35,12 @@ public class SampleDataFactory {
     @Autowired
     private VariantAnnotator variantAnnotator;
 
+    private final VariantFactory variantFactory;
     private final PedigreeFactory pedigreeFactory;
     private final GeneFactory geneFactory;
     
     public SampleDataFactory() {
+        variantFactory = new VariantFactory();
         pedigreeFactory = new PedigreeFactory();
         geneFactory = new GeneFactory();
     }
@@ -50,8 +48,8 @@ public class SampleDataFactory {
     public SampleData createSampleData(Path vcfFilePath, Path pedigreeFilePath) {
         logger.info("Creating sample data from VCF and PED files: {}, {}", vcfFilePath, pedigreeFilePath);
         
-        SampleData sampleData = parseSampleDataFromVcfFile(vcfFilePath);
-        sampleData.setVcfFilePath(vcfFilePath);
+        SampleData sampleData = createSampleDataFromVcf(vcfFilePath);
+        
         //Don't try and create the Genes before annotating the Variants otherwise you'll have a single gene with all the variants in it...
         List<Gene> geneList = geneFactory.createGenes(sampleData.getVariantEvaluations());
         sampleData.setGenes(geneList);
@@ -62,66 +60,27 @@ public class SampleDataFactory {
         return sampleData;
     }
 
-    private SampleData parseSampleDataFromVcfFile(Path vcfFile) {
-        VCFReader vcfReader = createVcfParser(vcfFile);
-        SampleData sampleData = createSampleDataFromVcf(vcfReader);
+    private SampleData createSampleDataFromVcf(Path vcfFilePath) {
+        logger.info("Creating SampleData from VCF");
         
-        List<Variant> variantList = createVariants(vcfReader);
-        //Now we've got all the basic bits of data sorted we'll need to fill-in the details needed for analysis
+        VCFReader vcfReader = variantFactory.createVcfReader(vcfFilePath);
+        List<Variant> variantList = variantFactory.createVariants(vcfReader);
+
+        SampleData sampleData = createSampleDataFromVcfMetaData(vcfReader);
         List<VariantEvaluation> variantEvaluations = createVariantEvaluations(variantList);
         sampleData.setVariantEvaluations(variantEvaluations);
-                
+        sampleData.setVcfFilePath(vcfFilePath);
+        
         return sampleData;
     }
-       
-    private VCFReader createVcfParser(Path vcfFile) {
-        VCFReader vcfReader = null;
-        try {
-            vcfReader = new VCFReader(vcfFile.toString());
-        } catch (VCFParseException ex) {
-            String message = String.format("Could not create VCFReader for VCF file: '%s'", vcfFile);
-            logger.error(message, ex);
-            throw new SampleDataCreationException(message, ex);
-        }
-        return vcfReader;
-    }
-    
-    private SampleData createSampleDataFromVcf(VCFReader vcfReader) {
-        logger.info("Creating SampleData from VCF");
 
-        try {
-            vcfReader.inputVCFheader();
-        } catch (VCFParseException ex) {
-            String message = String.format("Unable to parse header information from VCF file: '%s'", vcfReader.getVCFFileName());
-            logger.error(message, ex);
-            throw new SampleDataCreationException(message, ex);
-        }
-
+    private SampleData createSampleDataFromVcfMetaData(VCFReader vcfReader) {
         SampleData sampleData = new SampleData();
         sampleData.setVcfHeader(vcfReader.get_vcf_header());
         sampleData.setSampleNames(vcfReader.getSampleNames());
         sampleData.setNumberOfSamples(vcfReader.getNumberOfSamples());
 
         return sampleData;
-    }
-
-    private List<Variant> createVariants(VCFReader vcfReader) {
-        List<Variant> variantList = new ArrayList<>();
-        logger.info("Parsing Variants from VCF");
-        try {
-
-            Iterator<Variant> variantIterator = vcfReader.getVariantIterator();
-            while (variantIterator.hasNext()) {
-                variantList.add(variantIterator.next());
-
-            }
-        } catch (JannovarException ex) {
-            String message = String.format("Error parsing Variants from VCF file '%s'", vcfReader.getVCFFileName());
-            logger.error(message, ex);
-            throw new SampleDataCreationException(message, ex);
-        }
-
-        return variantList;
     }
 
     private List<VariantEvaluation> createVariantEvaluations(List<Variant> variantList) {
@@ -138,10 +97,4 @@ public class SampleDataFactory {
         return variantEvaluations;
     }
 
-    public class SampleDataCreationException extends RuntimeException {
-
-        public SampleDataCreationException(String format, Exception e) {
-            super(format, e);
-        }
-    }
 }