+++ exomiser-core/src/main/java/de/charite/compbio/exomiser/core/factories/VariantAnnotator.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package de.charite.compbio.exomiser.core.util;
+package de.charite.compbio.exomiser.core.factories;
 
 import jannovar.annotation.AnnotationList;
 import jannovar.exception.AnnotationException;