+++ exomiser-core/src/main/java/de/charite/compbio/exomiser/core/writer/VcfResultsWriter.java
@@ -25,6 +25,7 @@ import org.slf4j.LoggerFactory;
 
 /**
  *
+ * @see http://samtools.github.io/hts-specs/VCFv4.1.pdf
  * @author Jules Jacobsen <jules.jacobsen@sanger.ac.uk>
  */
 public class VcfResultsWriter implements ResultsWriter {
@@ -56,62 +57,102 @@ public class VcfResultsWriter implements ResultsWriter {
     @Override
     public String writeString(SampleData sampleData, ExomiserSettings settings, List<Priority> priorityList) {
         StringBuilder stringBuilder = new StringBuilder();
+        addHeader(stringBuilder);
+        addUnAnnotatedVariants(sampleData, stringBuilder);
+        addVariantsByGene(sampleData, stringBuilder);
+        return stringBuilder.toString();
+    }
+
+    private void addHeader(StringBuilder stringBuilder) {
         stringBuilder.append(VCF_FILE_FORMAT_4_1_LINE);
-        //TODO: Add in the settings used and other data for Will and Orion here
-        //(Issue #26 https://bitbucket.org/exomiser/exomiser/issue/26/vcf-output-format-requirements)
         stringBuilder.append(VCF_COLUMN_LINE);
+    }
+
+    private void addUnAnnotatedVariants(SampleData sampleData, StringBuilder stringBuilder) {
+        //report any variants without Jannovar annotations first to alert users of the problem
+        for (VariantEvaluation variantEval : sampleData.getUnAnnotatedVariantEvaluations()) {
+            buildVariantLine(variantEval, stringBuilder, null);
+        }
+    }
+
+    private void addVariantsByGene(SampleData sampleData, StringBuilder stringBuilder) {
         //write in the results for each variant in each gene
         for (Gene gene : sampleData.getGenes()) {
             stringBuilder.append(buildGeneVariantsString(gene));
         }
-        return stringBuilder.toString();
     }
 
     protected String buildGeneVariantsString(Gene gene) {
-        StringBuilder sb = new StringBuilder();
-        for (VariantEvaluation ve : gene.getVariantEvaluations()) {
-            Variant v = ve.getVariant();
-            //CHROM\t
-            addColumnField(sb, v.get_chromosome_as_string());
-            //POS\t
-            addColumnField(sb, String.valueOf(v.get_position()));
-            //ID\t
-            addColumnField(sb, ".");
-            //REF\t
-            addColumnField(sb, v.get_ref());
-            //ALT\t
-            addColumnField(sb, v.get_alt());
-            //QUAL\t
-            addColumnField(sb, String.valueOf(v.getVariantPhredScore()));
-            //FILTER\t
-            if (ve.passedFilters()) {
-                addColumnField(sb, "PASS");            
-            } else {
-                addColumnField(sb, formatFailedFilters(ve.getFailedFilterTypes()));
-            }
-            //INFO\t
-            String infoField = String.format("%s;EXOMISER_GENE=%s;EXOMISER_VARIANT_SCORE=%s;EXOMISER_GENE_PHENO_SCORE=%s;EXOMISER_GENE_VARIANT_SCORE=%s;EXOMISER_GENE_COMBINED_SCORE=%s", ve.getVariant().get_info(), gene.getGeneSymbol(), ve.getVariantScore(), gene.getPriorityScore(), gene.getFilterScore(), gene.getCombinedScore());
-            addColumnField(sb, infoField);
-            //FORMAT\t
-            addColumnField(sb, "GT");
-            //GENOTYPE\n
-            sb.append(v.getGenotypeAsString()).append("\n");
+        StringBuilder stringBuilder = new StringBuilder();
+        for (VariantEvaluation variantEval : gene.getVariantEvaluations()) {
+            buildVariantLine(variantEval, stringBuilder, gene);
         }
-        return sb.toString();
+        return stringBuilder.toString();
+    }
+
+    private void buildVariantLine(VariantEvaluation varEval, StringBuilder sb, Gene gene) {
+        Variant variant = varEval.getVariant();
+        //CHROM\t
+        addColumnField(sb, variant.get_chromosome_as_string());
+        //POS\t
+        addColumnField(sb, String.valueOf(variant.get_position()));
+        //ID\t
+        addColumnField(sb, ".");
+        //REF\t
+        addColumnField(sb, variant.get_ref());
+        //ALT\t
+        addColumnField(sb, variant.get_alt());
+        //QUAL\t
+        addColumnField(sb, String.valueOf(variant.getVariantPhredScore()));
+        //FILTER\t
+        addColumnField(sb, makeFiltersField(varEval));
+        //INFO\t
+        addColumnField(sb, makeInfoField(varEval, gene));
+        //FORMAT\t
+        addColumnField(sb, "GT");
+        //GENOTYPE\n
+        sb.append(variant.getGenotypeAsString()).append("\n");
     }
 
     protected void addColumnField(StringBuilder sb, String value) {
         sb.append(value).append(VCF_COLUMN_DELIMITER);
     }
 
-    protected String formatFailedFilters(Set<FilterType> failedFilters) {
-        StringBuilder stringBuilder = new StringBuilder();
-        for (FilterType filterType : failedFilters) {
-            stringBuilder.append(filterType.toString()).append(";");
+    protected String makeFiltersField(VariantEvaluation variantEvaluation) {
+        switch(variantEvaluation.getFilterStatus()) {
+            case FAILED:
+                return makeFailedFilters(variantEvaluation.getFailedFilterTypes());
+            case PASSED:
+                return "PASS";
+            case UNFILTERED:
+                return "";
+            default:
+                return "";
         }
-        //remove the final semi-colon
-        int sbLength = stringBuilder.length();
-        return stringBuilder.substring(0, sbLength - 1);
     }
 
+    protected String makeFailedFilters(Set<FilterType> failedFilterTypes) {
+        StringBuilder stringBuilder = new StringBuilder();
+        for (FilterType filterType : failedFilterTypes) {
+            stringBuilder.append(filterType.toString()).append(";");
+        }
+        removeTrailingSemiColon(stringBuilder);
+        return stringBuilder.toString();
+    }
+
+    private void removeTrailingSemiColon(StringBuilder stringBuilder) {
+        if (stringBuilder.charAt(stringBuilder.length() - 1) == ';') {
+            stringBuilder.deleteCharAt(stringBuilder.length() - 1);
+        }
+    }
+
+    //TODO: Add in the settings used and other data for Will and Orion here
+    //(Issue #26 https://bitbucket.org/exomiser/exomiser/issue/26/vcf-output-format-requirements)
+    private String makeInfoField(VariantEvaluation ve, Gene gene) {
+        String existingVcfInfoField = ve.getVariant().get_info();
+        if (ve.hasAnnotations() && gene != null) {
+            return String.format("%s;EXOMISER_GENE=%s;EXOMISER_VARIANT_SCORE=%s;EXOMISER_GENE_PHENO_SCORE=%s;EXOMISER_GENE_VARIANT_SCORE=%s;EXOMISER_GENE_COMBINED_SCORE=%s", existingVcfInfoField, gene.getGeneSymbol(), ve.getVariantScore(), gene.getPriorityScore(), gene.getFilterScore(), gene.getCombinedScore());
+        }
+        return String.format("%s;VARIANT NOT ANALYSED - NO GENE ANNOTATIONS", existingVcfInfoField);
+    }
 }