+++ exomiser-core/src/test/java/de/charite/compbio/exomiser/core/model/VariantEvaluationTest.java
@@ -6,7 +6,6 @@
 package de.charite.compbio.exomiser.core.model;
 
 import de.charite.compbio.exomiser.core.filter.FilterResult;
-import de.charite.compbio.exomiser.core.filter.FilterResultStatus;
 import static de.charite.compbio.exomiser.core.filter.FilterResultStatus.FAIL;
 import static de.charite.compbio.exomiser.core.filter.FilterResultStatus.PASS;
 import de.charite.compbio.exomiser.core.filter.FilterType;
@@ -40,16 +39,17 @@ import org.mockito.runners.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class VariantEvaluationTest {
 
-    VariantEvaluation instance;
+    private VariantEvaluation instance;
+    
     private static final Integer QUALITY = 2;
     private static final Integer READ_DEPTH = 6;
     private static final Genotype HETEROZYGOUS = Genotype.HETEROZYGOUS;
     private static final String GENE1_GENE_SYMBOL = "GENE1";
     private static final int GENE1_ENTREZ_GENE_ID = 1234567;
-    
+
     private static final String GENE2_GENE_SYMBOL = "GENE2";
     private static final int GENE2_ENTREZ_GENE_ID = 7654321;
-    
+
     private static final FilterResult FAIL_FREQUENCY_RESULT = new FrequencyFilterResult(0.1f, FAIL);
     private static final FilterResult PASS_FREQUENCY_RESULT = new FrequencyFilterResult(0.1f, PASS);
 
@@ -57,13 +57,16 @@ public class VariantEvaluationTest {
 
     @Mock
     Variant variant;
-    
+
     @Mock
     Variant variantInTwoGeneRegions;
-    
+
     @Mock
     Variant variantWithNullGeneSymbol;
-    
+
+    @Mock
+    Variant unAnnotatedVariant;
+
     @Before
     public void setUp() {
         GenotypeCall genotypeCall = new GenotypeCall(HETEROZYGOUS, QUALITY, READ_DEPTH);
@@ -77,7 +80,8 @@ public class VariantEvaluationTest {
         Mockito.when(variant.getGenotype()).thenReturn(genotypeCall);
         Mockito.when(variant.getVariantPhredScore()).thenReturn(2.2f);
         Mockito.when(variant.getVariantReadDepth()).thenReturn(READ_DEPTH);
-        
+        Mockito.when(variant.getAnnotation()).thenReturn("variant annotations...");
+
         Mockito.when(variantInTwoGeneRegions.getGeneSymbol()).thenReturn(GENE2_GENE_SYMBOL + "," + GENE1_GENE_SYMBOL);
         Mockito.when(variantInTwoGeneRegions.getEntrezGeneID()).thenReturn(GENE2_ENTREZ_GENE_ID);
         Mockito.when(variantInTwoGeneRegions.getChromosomeAsByte()).thenReturn((byte) 1);
@@ -87,9 +91,12 @@ public class VariantEvaluationTest {
         Mockito.when(variantInTwoGeneRegions.getGenotype()).thenReturn(genotypeCall);
         Mockito.when(variantInTwoGeneRegions.getVariantPhredScore()).thenReturn(2.2f);
         Mockito.when(variantInTwoGeneRegions.getVariantReadDepth()).thenReturn(READ_DEPTH);
-        
+
         Mockito.when(variantWithNullGeneSymbol.getGeneSymbol()).thenReturn(null);
-        
+
+        //This is hard-coding Jannovar's return values be aware this could change
+        Mockito.when(unAnnotatedVariant.getAnnotation()).thenReturn(".");
+
         instance = new VariantEvaluation(variant);
     }
 
@@ -119,7 +126,7 @@ public class VariantEvaluationTest {
         instance = new VariantEvaluation(variantInTwoGeneRegions);
         assertThat(instance.getGeneSymbol(), equalTo(GENE2_GENE_SYMBOL));
     }
-    
+
     @Test
     public void testGetGeneSymbolReturnsADotIfVariantReturnsANull() {
         instance = new VariantEvaluation(variantWithNullGeneSymbol);
@@ -130,7 +137,7 @@ public class VariantEvaluationTest {
     public void canGetEntrezGeneID() {
         assertThat(instance.getEntrezGeneID(), equalTo(GENE1_ENTREZ_GENE_ID));
     }
-    
+
     @Test
     public void testGetVariantStartPosition() {
         assertThat(instance.getVariantStartPosition(), equalTo(1));
@@ -186,7 +193,7 @@ public class VariantEvaluationTest {
         assertThat(instance.getFilterResults().size(), equalTo(2));
         assertThat(instance.getVariantScore(), equalTo(expectedScore));
     }
-    
+
     @Test
     public void testThatAddingPassAndFailFilterResultsUpdatesVariantScore() {
 
@@ -211,7 +218,7 @@ public class VariantEvaluationTest {
         instance.addFilterResult(FAIL_FREQUENCY_RESULT);
         assertThat(instance.getFailedFilterTypes(), equalTo(expectedFilters));
     }
-    
+
     @Test
     public void testGetFailedFilterTypesDontContainPassedFilterTypes() {
         Set<FilterType> expectedFilters = EnumSet.of(FAIL_FREQUENCY_RESULT.getFilterType());
@@ -221,11 +228,11 @@ public class VariantEvaluationTest {
 
         assertThat(instance.getFailedFilterTypes(), equalTo(expectedFilters));
     }
-    
+
     @Test
     public void testFailsFiltersWhenFailedFilterResultAdded() {
         instance.addFilterResult(FAIL_FREQUENCY_RESULT);
-        
+
         assertThat(instance.passedFilters(), is(false));
     }
 
@@ -260,6 +267,48 @@ public class VariantEvaluationTest {
     }
 
     @Test
+    public void testPassesFiltersWhenOnlyPassedFiltersHaveBeenApplied() {
+        instance.addFilterResult(PASS_QUALITY_RESULT);
+        instance.addFilterResult(PASS_FREQUENCY_RESULT);
+        assertThat(instance.getFailedFilterTypes().isEmpty(), is(true));
+        assertThat(instance.getFilterResults().isEmpty(), is(false));
+        assertThat(instance.passedFilters(), is(true));
+    }
+
+    @Test
+    public void testFailsFiltersWhenPassedAndFailedFiltersHaveBeenApplied() {
+        instance.addFilterResult(PASS_QUALITY_RESULT);
+        instance.addFilterResult(FAIL_FREQUENCY_RESULT);
+        assertThat(instance.getFailedFilterTypes().isEmpty(), is(false));
+        assertThat(instance.getFilterResults().isEmpty(), is(false));
+        assertThat(instance.passedFilters(), is(false));
+    }
+
+    @Test
+    public void testFilterStatusWhenNoFiltersHaveBeenApplied() {
+        assertThat(instance.getFilterStatus(), equalTo(FilterStatus.UNFILTERED));
+    }
+
+    @Test
+    public void testFilterStatusWhenFiltersHaveBeenAppliedAndPassed() {
+        instance.addFilterResult(PASS_QUALITY_RESULT);
+        assertThat(instance.getFilterStatus(), equalTo(FilterStatus.PASSED));
+    }
+
+    @Test
+    public void testFilterStatusWhenFiltersHaveBeenAppliedAndFailed() {
+        instance.addFilterResult(FAIL_FREQUENCY_RESULT);
+        assertThat(instance.getFilterStatus(), equalTo(FilterStatus.FAILED));
+    }
+
+    @Test
+    public void testFilterStatusWhenFiltersHaveBeenAppliedWithPassAndFailedResults() {
+        instance.addFilterResult(PASS_QUALITY_RESULT);
+        instance.addFilterResult(FAIL_FREQUENCY_RESULT);
+        assertThat(instance.getFilterStatus(), equalTo(FilterStatus.FAILED));
+    }
+
+    @Test
     public void testPassesFilterIsTrueWhenPassedFilterResultAdded() {
         FilterType passedFilterType = PASS_QUALITY_RESULT.getFilterType();
 
@@ -286,7 +335,7 @@ public class VariantEvaluationTest {
 
         assertThat(instance.getFilterResult(filterType), nullValue());
     }
-    
+
     @Test
     public void testGetFilterResultOfPassedFilter() {
         FilterType filterType = PASS_FREQUENCY_RESULT.getFilterType();
@@ -294,5 +343,16 @@ public class VariantEvaluationTest {
         instance.addFilterResult(PASS_FREQUENCY_RESULT);
 
         assertThat(instance.getFilterResult(filterType), equalTo(PASS_FREQUENCY_RESULT));
+    }
+
+    @Test
+    public void testHasAnnotationsIsFalse() {
+        VariantEvaluation varEvalWithNoVariantAnnotations = new VariantEvaluation(unAnnotatedVariant);
+        assertThat(varEvalWithNoVariantAnnotations.hasAnnotations(), is(false));
+    }
+
+    @Test
+    public void testHasAnnotationsIsTrue() {
+        assertThat(instance.hasAnnotations(), is(true));
     }
 }